Add macOS CMake build support
Port the Visual Studio project to build on macOS with CMake. Key changes:
- Add CMakeLists.txt using Homebrew SDL2/SDL2_image and system OpenGL
- Add source/compat.h shim for MSVC-specific APIs (fopen_s, errno_t, sscanf_s)
- Fix bctypes.h u32/i32 to use int instead of long (8 bytes on arm64 LP64)
- Fix SDL_Surface.h -> SDL_surface.h casing for case-sensitive filesystems
- Replace lzsa lib.h include with direct defines to avoid C++ enum forward ref
- Use nfd_cocoa.m on macOS, nfd_win.cpp on Windows for native file dialogs
- Fix Windows backslash paths to forward slashes (cross-platform)
- Add platform guard for dirent.h (system header on macOS, bundled on Windows)

+++ source/common/bctypes.h
@@ -11,8 +11,8 @@ typedef	signed char			i8;
 typedef unsigned char		u8;
 typedef signed short		i16;
 typedef unsigned short		u16;
-typedef signed long			i32;
-typedef unsigned long		u32;
+typedef signed int			i32;
+typedef unsigned int		u32;
 
 typedef signed long long	i64;
 typedef unsigned long long	u64;
